WARNING: Code is broken
TWEANNLink and LinkGene got mixed up somewhere. Maybe. Tracing this tomorrow

+++ Assets/Scripts/NEAT/TWEANN/ActivationFunctions/ActivationFunctions.cs
@@ -30,7 +30,7 @@ public class ActivationFunctions  {
     /// <param name="fType">Enum <see cref="FTYPE" /></param>
     /// <param name="sum">Input sent to node</param>
     /// <returns></returns>
-    public static double Activation(FTYPE fType, double sum)
+    public static float Activation(FTYPE fType, float sum)
     {
         return activationFunctions[fType].Function(sum);
     }
