+++ Assets/Scripts/NEAT/TWEANN/TWEANNGenotype/LinkGene.cs
@@ -5,10 +5,10 @@ using UnityEngine;
 public class LinkGene : Gene {
 
     protected long sourceInnovation, targetInnovation;
-    protected double weight;
+    protected float weight;
     protected bool active;
 
-    public LinkGene(long sourceInnovation, long targetInnovation, double weight, long innovation) : base(innovation)
+    public LinkGene(long sourceInnovation, long targetInnovation, float weight, long innovation) : base(innovation)
     {
         this.sourceInnovation = sourceInnovation;
         this.targetInnovation = targetInnovation;
@@ -31,12 +31,12 @@ public class LinkGene : Gene {
         return false;
     }
 
-    public double GetWeight()
+    public float GetWeight()
     {
         return weight;
     }
 
-    public void SetWeight(double weight)
+    public void SetWeight(float weight)
     {
         this.weight = weight;
     }
